Consolidate rules for determining if user's email has been verified
The code to determine if a user has verified their email got longer b/c we need a reliable way to check if the meta data exists at all. As a result, needed some refactoring and reuse.

The needs_validation() function now returns a boolean false if:

1. There is no meta key called "verify-lock" for the given user.
2. The "verify-lock" meta key has a value of "unlocked".

Otherwise, it will return the validation hash stored for that user.