Use a CSPRNG to generate the hash
Use `random_bytes( 16 )` to get 128-bits of entropy via a CSPRNG.

See the advice here: https://security.stackexchange.com/questions/64797/not-hashing-email-registration-confirmation-link#comment105340_64847

That whole thread is helpful in understanding this change.

It's probably not a *huge* security risk, but unless it's a *cryptographically-secure* pseudo-random number generator, the door is left open to users with unverified emails gaining access. The whole point of this plugin is to avoid that, so this change should be an improvement.